Alter and refactor the unenrol_student_from_module method
Altered the method so that the user only has the option of unenrolling
a student from an optional module on their current academic year. This
then made it possible to refactor the code further. Also added the
options for the user to unenrol multiple times or switch to enrolling,
for the same student.